backend/auth-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.8.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
requests-oauthlib==1.3.1
alembic==1.13.1
redis==5.0.1
celery==5.3.4
